Python/3_python_api_dev/1.backups_and_course_source/backup_5_at_022/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
passlib
bcrypt==4.0.1
